Moved all libraries into one file #11
