+++ libraries/azure_rg.rb
@@ -1,5 +1,5 @@
 
-require File.expand_path(File.dirname(__FILE__) + '/common/helpers')
+require 'azure_backend'
 
 # Class to test the resources in Resource Groups
 #
